refactor: moved router and controller files to respective folders
