--- src/imageGenerate.route.js
@@ -1,14 +1,0 @@
-import express, { Router } from "express";
-
-const router = Router();
-
-// dummy route
-router.get("/", async (req, res) => {
-  return res.status(200).send({
-    message: "Response message from get",
-  });
-});
-
-// TODO: define actual routes
-
-export default router;
